Cleanup
basically just remove old unused code
