+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Auton/ActorTest.java
@@ -7,8 +7,11 @@ import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
+import org.firstinspires.ftc.teamcode.Control.Actor.Actor;
+import org.firstinspires.ftc.teamcode.Control.Actor.ClawAction;
+import org.firstinspires.ftc.teamcode.Control.Actor.LiftAction;
+import org.firstinspires.ftc.teamcode.Control.Actor.MvntAction;
 import org.firstinspires.ftc.teamcode.Control.Movement;
-import org.firstinspires.ftc.teamcode.Subsystems.Claw;
 import org.firstinspires.ftc.teamcode.Subsystems.Robot;
 import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
