+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Subsystems/Drive.java
@@ -6,7 +6,6 @@ import com.qualcomm.robotcore.hardware.DcMotorEx;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 import com.qualcomm.robotcore.hardware.IMU;
 
-import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 import org.firstinspires.ftc.teamcode.Auton.Config;
 
 public class Drive {
@@ -74,11 +73,6 @@ public class Drive {
         frontRight.setPower(frontRightPower);
         backLeft.setPower(backLeftPower);
         backRight.setPower(backRightPower);
-    }
-
-    // TEMP
-    public double getIMU() {
-        return imu.getRobotYawPitchRollAngles().getYaw(AngleUnit.DEGREES);
     }
 
     public static class DrivePowers {
