+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Subsystems/Hang.java
@@ -1,14 +1,10 @@
 package org.firstinspires.ftc.teamcode.Subsystems;
 
-import com.qualcomm.robotcore.hardware.CRServo;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
 import com.qualcomm.robotcore.hardware.Gamepad;
 import com.qualcomm.robotcore.hardware.HardwareMap;
-import com.qualcomm.robotcore.hardware.Servo;
-import com.qualcomm.robotcore.util.ElapsedTime;
 
-import org.firstinspires.ftc.robotcore.internal.system.CloseableOnFinalize;
 import org.firstinspires.ftc.teamcode.Auton.Config;
 
 public class Hang {
@@ -36,6 +32,7 @@ public class Hang {
     }
 
     public void hangNow() {
+        //noinspection InfiniteLoopStatement
         for (;;) {
             hangMotor.setPower(Config.hangPower);
         }
