+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/DriveConstants.java
@@ -93,11 +93,6 @@ public class DriveConstants {
     public static double MAX_ANG_VEL = Math.toRadians(246.50000000000003);
     public static double MAX_ANG_ACCEL = Math.toRadians(246.50000000000003);
 
-//    public static double MAX_VEL = 20;
-//    public static double MAX_ACCEL = 15;
-//    public static double MAX_ANG_VEL = Math.toRadians(30);
-//    public static double MAX_ANG_ACCEL = Math.toRadians(30);
-
     /*
      * Adjust the orientations here to match your robot. See the FTC SDK documentation for details.
      */
