--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/opmodes/OdomTest.java
@@ -1,15 +1,0 @@
-//package org.firstinspires.ftc.teamcode.drive.opmodes;
-//
-//import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-//import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-//
-//@TeleOp(name = "Mecanum Drive", group = "Linear Opmode")
-//public class OdomTest extends LinearOpMode {
-//
-//    @Override
-//    public void runOpMode() {
-//
-//        EnccenterOdometer = new MotorEx(hardwareMap, "frontRight").encoder;
-//    }
-//}
-//
