+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/opmodes/TeleOpControl.java
@@ -8,16 +8,9 @@ import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
-import org.firstinspires.ftc.teamcode.AprilTags.AprilTagsDetectionPipeline;
-import org.firstinspires.ftc.teamcode.AprilTags.AprilTagsInit;
 import org.firstinspires.ftc.teamcode.Auton.Config;
-import org.firstinspires.ftc.teamcode.Controllers.MotionProfile;
 import org.firstinspires.ftc.teamcode.Subsystems.*;
 import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
-import org.firstinspires.ftc.teamcode.drive.TwoWheelTrackingLocalizer;
-import org.firstinspires.ftc.teamcode.util.DashboardUtil;
-
 
 @TeleOp(name = "Mecanum Drive", group = "Linear Opmode")
 public class TeleOpControl extends LinearOpMode {
@@ -41,7 +34,6 @@ public class TeleOpControl extends LinearOpMode {
         boolean lastX = false;
 
         waitForStart();
-        robot.odom.reset();
         robot.drive.imu.resetYaw();
         timer.reset();
 
@@ -49,17 +41,7 @@ public class TeleOpControl extends LinearOpMode {
             drive.updatePoseEstimate();
             Pose2d poseEstimate = drive.getPose();
 
-//            telemetry.addData("left x: ", gamepad1.left_stick_x);
-//            telemetry.addData("left y: ", gamepad1.left_stick_y);
-//            telemetry.addData("right x: ", gamepad1.right_stick_x);
-//            telemetry.addData("right y: ", gamepad1.right_stick_y);
-
             robot.lift.liftTeleOp(gamepad1, tel); // LIFT
-//            telemetry.addData("kill time", robot.lift.timer.milliseconds());
-//            telemetry.addData("hasReset", robot.lift.hasResetKill);
-//            telemetry.addData("macro", robot.lift.currentMode);
-//            telemetry.addData("hold", robot.lift.holdingPos);
-//            telemetry.addData("kill", robot.lift.startedKill);
             robot.claw.input(gamepad1);
             robot.hang.input(gamepad1);
 //            robot.claw.setPower(-1, -1);
@@ -74,22 +56,16 @@ public class TeleOpControl extends LinearOpMode {
                 double turn = gamepad1.right_stick_x;
                 double multiplier = slowMode ? 0.3 : 1;
                 robot.drive.mecanumDrive(power * multiplier, strafe * multiplier, turn * multiplier);
-//                robot.drive.mecanumDrive(power, strafe, turn);
-//                telemetry.addData("drive: ", "robotCentric");
             } else {
                 double elapsed = timer.seconds();
 
                 double y = -gamepad1.left_stick_y;
                 double x = gamepad1.left_stick_x;
-                double h = gamepad1.right_stick_x;
-
-                double profileY = MotionProfile.motion_profile(Odometry.MAX_ACCEL, Odometry.MAX_VELOCITY, y, elapsed);
-                double profileX = MotionProfile.motion_profile(Odometry.MAX_ACCEL, Odometry.MAX_VELOCITY, x, elapsed);
-                double rx = MotionProfile.motion_profile(Odometry.MAX_ACCEL * 4, Odometry.MAX_VELOCITY * 4, h, elapsed);
+                double rx = gamepad1.right_stick_x;
 
                 double botHeading = -poseEstimate.getHeading();
-                double rotX = profileX * Math.cos(-botHeading) - profileY * Math.sin(-botHeading);
-                double rotY = profileX * Math.sin(-botHeading) + profileY * Math.cos(-botHeading);
+                double rotX = x * Math.cos(-botHeading) - y * Math.sin(-botHeading);
+                double rotY = x * Math.sin(-botHeading) + y * Math.cos(-botHeading);
 
                 rotX = rotX * 1.1;
 
@@ -107,14 +83,9 @@ public class TeleOpControl extends LinearOpMode {
             tel.addData("drive", Config.fieldCentric);
 //            tel.addData("TEMP", apriltags.camera.getWhiteBalanceControl().getWhiteBalanceTemperature());
             tel.addData("pose", poseEstimate);
-            tel.addData("!LEFT ENCODER: ", robot.odom.getEncoders()[0]);
-            tel.addData("!RIGHT ENCODER: ", robot.odom.getEncoders()[1]);
-            tel.addData("!CENTER ENCODER: ", robot.odom.getEncoders()[2]);
-            tel.addData("lift owo: ", robot.lift.leftLift.getCurrentPosition());;
+            tel.addData("lift: ", robot.lift.leftLift.getCurrentPosition());
             tel.update();
 
-//            telemetry.addData("maxfps", apriltags.camera.getCurrentPipelineMaxFps());
-//            telemetry.addData("fps", apriltags.camera.getFps());
             telemetry.addData("SlowMode: ", slowMode);
             telemetry.addData("lift owo: ", robot.lift.leftLift.getCurrentPosition());
