+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/roadrunner_opmodes/LocalizationTest.java
@@ -27,13 +27,6 @@ public class LocalizationTest extends LinearOpMode {
         waitForStart();
 
         while (!isStopRequested()) {
-//            drive.setWeightedDrivePower(
-//                    new Pose2d(
-//                            -gamepad1.left_stick_y,
-//                            gamepad1.left_stick_x,
-//                            gamepad1.right_stick_x
-//                    )
-//            );
             double power = -gamepad1.left_stick_y; // remember this is reversed
             double strafe = gamepad1.left_stick_x * 1.1; // counteract imperfect strafing
             double turn = gamepad1.right_stick_x;
