+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/roadrunner_opmodes/StraightTest.java
@@ -41,6 +41,7 @@ public class StraightTest extends LinearOpMode {
         telemetry.addData("finalHeading", poseEstimate.getHeading());
         telemetry.update();
 
+        //noinspection StatementWithEmptyBody
         while (!isStopRequested() && opModeIsActive()) ;
     }
 }
